% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_env.R
\name{xp_checkBalance}
\alias{xp_checkBalance}
\title{检查司法存量的余量}
\usage{
xp_checkBalance(
  session = xp_session(),
  key_path = "data-raw/cfg",
  password = "396283"
)
}
\arguments{
\item{session}{会议}

\item{key_path}{key所在位置}

\item{password}{支付安全码}
}
\value{
返回值
}
\description{
检查司法存量的余量
}
\examples{
xp_checkBalance()
}
