% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_env.R
\name{xp_checkPrivateKey}
\alias{xp_checkPrivateKey}
\title{检查private.key是否存在}
\usage{
xp_checkPrivateKey(session = xp_session(), key_path = "data-raw/cfg")
}
\arguments{
\item{session}{会话管理器}

\item{key_path}{key所在的路径}
}
\value{
返回值
}
\description{
检查private.key是否存在
}
\examples{
xp_checkPrivateKey()
}
