% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_certMngr.R
\name{xp_queryCert}
\alias{xp_queryCert}
\title{查询证书}
\usage{
xp_queryCert(
  session = xp_session(),
  key_path = "data-raw/cfg",
  txtId = "5974108e7624f1f8c955ad170eeb84862c6322cb295a6ab354b4f4d6323c135d",
  password = "123456"
)
}
\arguments{
\item{session}{会话}

\item{key_path}{privatekey}

\item{txtId}{文件ID}

\item{password}{文件名}
}
\value{
返回值
}
\description{
查询证书
}
\examples{
xp_downloadCert()
}
