% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_file.R
\name{xp_QueryFile}
\alias{xp_QueryFile}
\title{查询可信存证}
\usage{
xp_QueryFile(
  session = xp_session(),
  key_path = "data-raw/cfg",
  password = "396283",
  txtId = "5974108e7624f1f8c955ad170eeb84862c6322cb295a6ab354b4f4d6323c135d"
)
}
\arguments{
\item{session}{会话管理员}

\item{key_path}{private.key所在目录}

\item{password}{安全码}

\item{txtId}{文件类型}
}
\value{
返回值
}
\description{
查询可信存证
}
\examples{
xp_upload()
}
